# src/main/resources/static/swagger.yml
openapi: 3.0.3
info:
  title: Online Shop
  description: |-
    Aplicación que se encarga de habilitar endpoints para el consumo REST de la base de datos en la tienda online.
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
...
paths:
...
  /shop/create-product:
    post:
      tags:
        - products
      summary: Crear un producto
      description: Crear un producto
      operationId: createProduct
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '500':
          description: Internal Server Error - IOException occurred
  /shop/update-product/{productId}:
    patch:
      tags:
        - products
      summary: Actualizar un producto
      description: Actualizar un producto
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: Id del producto
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
        '500':
          description: Internal Server Error - IOException occurred
  /shop/product/{productId}:
    get:
      tags:
        - products
      summary: Encontrar un producto por ID
      description: Econtrar un producto por su ID correspondiente.
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Product not found
  /shop/product/delete-product/{productId}:
    delete:
      tags:
        - products
      summary: Eliminar un producto por ID
      description: Eliminar un producto por su ID correspondiente.
      operationId: deleteProductById
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Product not found
...
  /shop/new-purchase:
    post:
      tags:
        - purchase
      summary: Hacer un pedido
      description: Hacer un pedido en la tienda online
      operationId: newPurchase
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseRequest'

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
  /shop/find-purchase/{purchaseId}:
    get:
      tags:
        - purchase
      summary: Encontrar un pedido por ID
      description: Econtrar un pedido por su ID correspondiente.
      operationId: getPurchaseById
      parameters:
        - name: purchaseId
          in: path
          description: ID del pedido
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Purchase not found
  /shop/remove-purchase/{purchaseId}:
    delete:
      tags:
        - purchase
      summary: Cancelar pedido
      description: Cancelar un pedido por su ID
      operationId: deletePurchaseById
      parameters:
        - name: purchaseId
          in: path
          description: ID del pedido a cancelar
          required: true
          schema:
            type: integer
            format: int64

      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
                example: Purchase canceled successfully
        '400':
          description: Invalid ID supplied
        '404':
          description: Purchase not found
  /shop/purchase-cancellation/{purchaseId}:
    patch:
      tags:
        - purchase
      summary: Peticion al administrador de cancelar un pedido por parte del usuario
      description: Cambia el estado del pedido a cancellation_pending
      operationId: purchaseCancellationRequest
      parameters:
        - name: purchaseId
          in: path
          description: ID del pedido a cancelar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Purchase not found
  /shop/purchase-cancellation-validation/{purchaseId}:
    patch:
      tags:
        - purchase
      summary: El administrador cancela el pedido al que se ha hecho peticion
      description: Cambia el estado del pedido a cancellation
      operationId: purchaseCancellationValidation
      parameters:
        - name: purchaseId
          in: path
          description: ID del pedido a cancelar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Purchase not found
...
  /users/{username}:
    get:
      tags:
        - users
      summary: Obtener usuario por nombre
      description: ''
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: 'Nombre del usuario a buscar '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
  /users/delete/{username}:
    patch:
      tags:
        - users
      summary: Desactivar usuario
      description: Desactivar acceso del usuario
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: Nombre de usuario
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
  /users/update-user:
    patch:
      tags:
        - users
      summary: Actualizar usuario
      description: Actualizar datos del usuario usuario.
      operationId: updateUser
      requestBody:
        description: Datos del usuario
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid username supplied
        '404':
          description: User not found
...
components:
  schemas:
...
    CartDetailsRequest:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 2
        quantity:
          type: integer
          example: 1
...
    ProductRequest:
      type: object
      properties:
        name:
          type: string
          example: name
        description:
          type: string
          example: descripcion del producto
        size:
          type: string
          example: "24 x 24"
        type:
          type: string
          example: cuadro
        price:
          type: number
          example: 20.99
        image:
          type: string
          format: binary
          example: photo.jpg
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: name
        description:
          type: string
          example: descripcion del producto
        size:
          type: string
          example: "24 x 24"
        type:
          type: string
          example: cuadro
        price:
          type: number
          format: float
          example: 20.99
        image:
          type: string
          example: photo.jpg
    PurchaseRequest:
      type: object
      properties:
        shippingAddress:
          type: string
    PurchaseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 2
        purchaseDate:
          type: string
          format: date-time
          example: 2024-05-26 timestampt
        estimatedDeliveryTime:
          type: string
          format: date
          example: "2024-05-29"
          description: "Fecha y hora estimada de entrega del pedido."
        totalAmount:
          type: number
          example: 29.99
        shippingAddress:
          type: string
        status:
          type: string
          description: Purchase Status
          example: approved
          enum:
            - placed
            - approved
            - delivered
            - finished
            - cancellation_pending
            - cancelled
        cartDetails:
          type: array
          items:
            $ref: '#/components/schemas/CartDetailsRequest'
...
    CustomerRequest:
      type: object
      properties:
        name:
          type: string
          example: John
        password:
          type: string
          example: '12345'
        surname:
          type: string
          example: James
        surname2:
          type: string
          example: James
        address:
          type: string
          example: Calle nombre n3 1b
        city:
          type: string
          example: Chiclana de la Frontera
        province:
          type: string
          example: Cadiz
        region:
          type: string
          example: España
        postalCode:
          type: string
          example: "11130"
        email:
          type: string
          example: john@email.com
        phone:
          type: string
          example: '12345'
        status:
          type: boolean
          description: Estado del usuario
          example: true
...
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: theUser
        password:
          type: string
          example: '12345'
        name:
          type: string
          example: John
        surname:
          type: string
          example: James
        surname2:
          type: string
          example: James
        address:
          type: string
          example: Calle nombre n3 1b
        city:
          type: string
          example: Chiclana de la Frontera
        province:
          type: string
          example: Cadiz
        region:
          type: string
          example: España
        postalCode:
          type: string
          example: "11130"
        email:
          type: string
          example: john@email.com
        phone:
          type: string
          example: '12345'
        status:
          type: boolean
          description: Estado del usuario
          example: true
        rol:
          type: string
          format: enum
            - USER
            - ADMIN
          description: Rol de usuario
          example: USER
...